Wrap ssh-askpass feature in functions
Add 3 new fucntions:
zsh-quickstart-enable-ssh-askpass-require
zsh-quickstart-disable-ssh-askpass-require
zsh-quickstart-check-for-ssh-askpass

These will now enable or disable the ability to have ssh ask for your
passphrase on the command line or via a gui app on Linux. The default is
to look for ssh-askpass as documented in the manpages for ssh-add.

Add a check for ssh-askpass and prompt to install it. Add a link to the
man page to give more info when they see this error on Linux.

Signed-off-by: Darren <10267761+dglynn@users.noreply.github.com>